pyproject.toml
[tool.poetry]
name = "agator-db"
version = "0.1.0"
description = ""
authors = ["Noah Kamara <github@login.noahkamara.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rethinkdb = "^2.4.8"
pydantic = "^1.8.1"
python-dotenv = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
